_archive/deletable_candidates_20260215_191454/Steer/requirements.txt
# Last updated: 2025-12-15
# Notes:
# - 범위 지정(>=, <)을 유지해서 macOS/conda 환경에서 설치 충돌을 줄입니다.
# - 로컬 GGUF LLM은 `llama-cpp-python` (MPS/Metal) 기준입니다.

# === 기반 패키지 (숫자 연산/데이터핸들링) ===
numpy>=1.26,<3
pandas>=2.2.3,<2.3
six>=1.16,<2

# === 파일 처리 & IO ===
openpyxl>=3.1.2,<3.2
fastparquet>=2024.5.0,<2025
olefile>=0.47,<0.48
python-docx>=1.1.0,<1.3
python-pptx>=1.0.0,<1.1
pdfminer.six>=20231228,<20240600
PyMuPDF>=1.26.4,<1.27
pdfplumber>=0.11.4,<0.12
soundfile>=0.12,<0.14
python-dotenv>=1.0.1,<1.1
Pillow>=10.4,<12   # photo 에이전트(EXIF 날짜 정리)
geopy>=2.4,<3   # photo 에이전트(GPS → 지명 변환)
imagehash>=4.3,<5   # photo 에이전트(중복 이미지 탐지)
mediapipe>=0.10,<1   # photo 에이전트(얼굴 감지)
insightface>=0.7,<1   # photo 에이전트(얼굴 인식/본인 구분)
onnxruntime>=1.16,<2   # insightface 의존성
ultralytics>=8.0,<9   # photo 에이전트(YOLO 객체 탐지)

# === 번역/텍스트 유틸 ===
deep-translator>=1.11.4,<2
rank-bm25>=0.2.2,<0.3

# === ML & AI 모델 ===
joblib>=1.4.2,<1.5
scikit-learn>=1.6.0,<1.7
sentence-transformers>=3.0.1,<4
faiss-cpu>=1.9,<2   # (Windows 3.13 → 실제 설치는 faiss-cpu==1.9.0.post1 권장)
faster-whisper>=1.0,<1.1   # 회의 비서 STT 백엔드(대체); 기본은 transformers 기반 wav2vec2
pyannote-audio>=3.1,<4   # 화자 분리 (Speaker Diarization)
transformers>=4.44,<5   # KoBART summarisation
llama-cpp-python>=0.3.16,<0.4   # 로컬 GGUF LLM (MPS/Metal 지원)
# pykospacing는 macOS/arm64 wheel 미제공 -> 필요시 수동설치
# kosspacing는 환경에 따라 배포/휠이 없어 설치가 실패할 수 있음 -> 필요 시 수동 설치
# kosspacing>=0.5,<0.6; python_version >= "3.12"   # torch 기반 띄어쓰기 대체
# py-hanspell은 구버전 pip API 의존 (필요 시 수동 설치)

# === 시스템 유틸 ===
watchdog>=4.0.0,<4.1
pywin32; platform_system == "Windows"
# pyhwp는 Python3 최신 버전 미지원 → 필요 시 별도 설치

# === 런타임 / CLI / API ===
tqdm>=4.66,<5
click>=8.1,<9
fastapi>=0.117,<0.118   # 테스트용 API 서버 의존성
httpx>=0.28,<0.29       # FastAPI TestClient 의존성
prefect>=2.19.0,<3
uvicorn>=0.30,<1
mlflow>=2.15,<3

# === UI (선택) ===
customtkinter>=5.2,<6
PySide6>=6.0.0
pyqtdarktheme>=0.1.7,<0.1.8   # Python 3.12에서는 0.1.x만 설치 가능(qdarktheme.load_stylesheet 사용)

# === Dev/Test ===
pytest>=8.0,<9
pytest-cov>=4.1.0
